Refactor vault folder by creating subfolders for different cloud providers
