+++ addons/resume/zz_desc_generated.go
@@ -1,0 +1,1 @@
+package resume
