+++ addons/vault/aws/aws_secrets_manager_test.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package vault
+package aws
 
 import (
 	"testing"
