+++ addons/vault/aws/zz_generated_doc.go
@@ -1,0 +1,1 @@
+package aws
