+++ pkg/apis/camel/v1/trait/zz_desc_generated.go
@@ -1,0 +1,1 @@
+package trait
