+++ pkg/apis/camel/v1/trait/zz_generated_doc.go
@@ -1,0 +1,1 @@
+package trait
